==> Final_Project/04_Visualization.Rmd <==
---
title: "Home Values Visualization"
author: "Ken Popkin"
---

```{r}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```

```{r echo = TRUE, message=FALSE, warning=FALSE}
library(dplyr)
library(ggplot2)
library(tidyverse)
library(reshape2)
library(ggpubr)
library(ggfortify)
library(psych)

require(scales)
```

### Introduction to Visualization
This is the 2nd of 3 notebooks that researches the question,<I> "Can education and wages be used as predictors for home values?". </I>  The first notebook completed data cleaning and transformation; this notebook will focus on visualization.

### About the Data
```{r echo=FALSE}
# load data
master <- read.csv('https://raw.githubusercontent.com/krpopkin/DATA607-Class-Repository/master/Final_Project/Data/master.csv')

cases = dim(master)[1]
features = dim(master)[2]
cat('Data for this project is', cases, 'rows and', features, 'features')
```

```{r data}
head(master,1)
```

Here is a helpful description of the 16 columns in the master data file.
```{r}
describe(master)
```

### Visualization
We'll start with a few histograms to get a feel for the normality of some of the key features.

<B> Histograms </B>  
Histograms of high school graduation rates and four year college graduation rates.   
```{r}
#Percent with high school diploma
hsd_hist <- ggplot(master, aes(x=hs_diploma_percent)) + geom_histogram()

#Percent with four year degrees
fyd_hist <- ggplot(master, aes(x=four_year_degree_percent)) + geom_histogram()

#Show education histograms
ggarrange(hsd_hist, fyd_hist, 
                    labels = c("hsd", "fyd"),
                    ncol = 2, nrow = 1,
                    common.legend = TRUE, legend = "bottom")
```
   
The high school diploma distribution appears normal, which makes sense given that a high school education is prevalent across the US and is publicly available.  The right skew to a four year degree reflects that even though university's are in every state, not everyone goes to (and completes) a four year degree for a number of reasons.

<I> Avg annual pay </I>  
```{r}
library(ggfortify)
ggplot(master, aes(x=avg_annual_pay)) +
  geom_histogram(bins = 50) + 
  scale_x_continuous(labels = comma) + 
  scale_y_continuous(labels = comma)
```
   
The histogram of average annual pay below is skewed right. This makes sense as all US states are in this dataset and there are some very high salaries in prestiguous counties in the US.

<I> Home values plot </I>  
```{r}
ggplot(master, aes(x=avg_home_value)) + 
  geom_histogram(bins = 50) + 
  scale_x_continuous(labels = comma) + 
  scale_y_continuous(labels = comma)
```
   
Similar to wages and for similar reasons, home values in some counties are very high, giving the distribution a right skew.

<B> Correlation Heat Maps </B>  
Now We'll move on to creating heat maps to highlight the correlations between Education, Wages, Resident Count and Home Values.  Our first heatmap will look at the four education categories of high school dropout, high school diploma, some college, and four year degree.  There should be high correlation between each of these so we'll select the one with highest correlation to home values to move forward with.

```{r education-heatmap}
#Select the data for the education heat map
ehm <- master %>%
  select(hs_dropout = dropout_percent, hs_diploma = hs_diploma_percent, some_college = some_college_percent,  college_degree = four_year_degree_percent, avg_home_value)

#Calculate the correlation values
cormat <- round(cor(ehm),2)

melted_cormat <- melt(cormat)

#Create the Heatmap
ggheatmap <- ggplot(data = melted_cormat, aes(Var2, Var1, fill = value))+
 geom_tile(color = "white")+
 scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
   midpoint = 0, limit = c(-1,1), space = "Lab", 
   name="Pearson\nCorrelation") +
  theme_minimal()+ 
 theme(axis.text.x = element_text(angle = 45, vjust = 1, 
    size = 12, hjust = 1))+
 coord_fixed()

reorder_cormat <- function(cormat){
# Use correlation between variables as distance
dd <- as.dist((1-cormat)/2)
hc <- hclust(dd)
cormat <-cormat[hc$order, hc$order]
}

ggheatmap + 
geom_text(aes(Var2, Var1, label = value), color = "black", size = 4)
```

The education heatmap correlation of 69% with a college degree and home values was somewhat expected. The -61% corelation for a high school diploma (so no higher education) with home values was initially surprising, but does make sense.

Next we'll generate a heatmap using wages, resident count, and home values.
```{r heatmap}
#Select the data for the heat map
chm <- master %>%
  select(avg_annual_pay, resident_count, avg_home_value)

#Calculate the correlation values
cormat <- round(cor(chm),2)

melted_cormat <- melt(cormat)

#Create the Heatmap
library(ggplot2)
ggheatmap <- ggplot(data = melted_cormat, aes(Var2, Var1, fill = value))+
 geom_tile(color = "white")+
 scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
   midpoint = 0, limit = c(-1,1), space = "Lab", 
   name="Pearson\nCorrelation") +
  theme_minimal()+ 
 theme(axis.text.x = element_text(angle = 45, vjust = 1, 
    size = 12, hjust = 1))+
 coord_fixed()

reorder_cormat <- function(cormat){
# Use correlation between variables as distance
dd <- as.dist((1-cormat)/2)
hc <- hclust(dd)
cormat <-cormat[hc$order, hc$order]
}

ggheatmap + 
geom_text(aes(Var2, Var1, label = value), color = "black", size = 4)
```

The 52% correlation for wages is strong, but not as strong as expected.  What's surprising is that average pay is showing a lower correlation than education in the previous heatmap.

<B> Scatter Charts </B>   
<I> Home Values Compared to Education, Wages, and Resident Count Colored by Census Region </I>   
There are 2,745 counties across 50 states, but only 4 census regions. We'll use regions to create several scatter charts of the relationship between education, wages, resident count, and home values.
```{r scatter-charts}
scatterdf <- select(master, region, region_num, hs_diploma_percent, four_year_degree_percent, avg_annual_pay, resident_count, avg_home_value)

#Scatter chart of hs diploma percent
hsd_scatter <- ggplot(scatterdf, aes(avg_home_value, hs_diploma_percent, colour = region)) + 
  geom_point(alpha=0.5, position = 'jitter') +
  scale_x_continuous(labels = comma) + 
  scale_y_continuous(labels = comma)

#Scatter chart of four year degree percent
fyd_scatter <- ggplot(scatterdf, aes(avg_home_value, four_year_degree_percent, colour = region)) + 
  geom_point(alpha=0.5, position = 'jitter') +
  scale_x_continuous(labels = comma) + 
  scale_y_continuous(labels = comma)

#Scatter chart of average annual pay
aap_scatter <- ggplot(scatterdf, aes(avg_home_value, avg_annual_pay, colour = region)) + 
  geom_point(alpha=0.5, position = 'jitter') +
  scale_x_continuous(labels = comma) + 
  scale_y_continuous(labels = comma)

#Scatter chart of resident count
rc_scatter <- ggplot(scatterdf, aes(avg_home_value, resident_count, colour = region)) + 
  geom_point(alpha=0.5, position = 'jitter') +
  scale_x_continuous(labels = comma) + 
  scale_y_continuous(labels = comma)

#Show scatter charts
ggarrange(hsd_scatter, fyd_scatter, aap_scatter, rc_scatter, 
                    labels = c("hsd", "fyd", "aap", "rc"),
                    ncol = 2, nrow = 2,
                    common.legend = TRUE, legend = "bottom")
```

<B> Violin Plot </B>   
We'll conclude this visuals section with comparisons using violin plots   
<I> Home Values </I>
```{r violin-plots}
#Violin of home values
hv_violin <- ggplot(scatterdf, aes(region_num, avg_home_value, colour = region)) + 
  geom_violin(trim = FALSE) 

#Display the violin plots
ggarrange(hv_violin,
          labels = c("hv"),
          ncol = 1, nrow = 1,
          common.legend = TRUE, legend = "top")
```

<I> Education </I>
```{r}
#Violin of high school diploma
hsd_violin <- ggplot(scatterdf, aes(region_num, hs_diploma_percent, colour = region)) + 
  geom_violin(trim = FALSE)

#Violin of four year degree
fyd_violin <- ggplot(scatterdf, aes(region_num, four_year_degree_percent, colour = region)) +
  geom_violin(trim = FALSE)

#Display the violin plots
ggarrange(hsd_violin, fyd_violin,
          labels = c("hsd", "fyd"),
          ncol = 2, nrow = 1,
          common.legend = TRUE, legend = "bottom")
```

<I> Wages and Resident Count </I>
```{r}
#Violin of wages
aap_violin <- ggplot(scatterdf, aes(region_num, avg_annual_pay, colour = region)) + 
  geom_violin(trim = FALSE) + 
  scale_y_continuous(labels = comma)

#Violin of resident count
rc_violin <- ggplot(scatterdf, aes(region_num, resident_count, colour = region)) + 
  geom_violin(trim = FALSE) + 
  scale_y_continuous(labels = comma)

#Display the violin plots
ggarrange(aap_violin, rc_violin,
          labels = c("aap", "rc"),
          ncol = 2, nrow = 1,
          common.legend = TRUE, legend = "bottom")
```

<B> Summary </B>   
The visualizations have served their purpose of making us much more familiar with the data.  In the final notebook (link to Analysis and Conclusions), we'll use the information learned here for further analysis; these visualizations make a good case for application of a regression model to this data.  

Here is a summary of what's been learned so far... <br>
1. High school education is normally distributed.<br><br>
2. Four year degrees, wages, and home values are all right skewed (this consistency in their skew could help a model's ability to make predictions.)<br><br>3. Home values do appear to be correlated to education and wages.<br><br>4. In the violin plot we also saw that resident count in each county might be a feature worth including in a model to predict home values.<br><br>5. Overall, viewing this data by census region worked well.  The visuals definitely highlight that the Midwest is the lowest cost area and that the West and Northeast are both higher cost and a source of most of the outlier data.  This is probably what most of us would have guessed, but it's always good to see data that validates what we've already perceived or been told is true.   

### References
[This article from STHDA](http://www.sthda.com/english/wiki/ggplot2-quick-correlation-matrix-heatmap-r-software-and-data-visualization) was a huge help in generating the heatmaps.

[This article from STHDA](http://www.sthda.com/english/articles/39-regression-model-diagnostics/161-linear-regression-assumptions-and-diagnostics-in-r-essentials/#diagnostic-plots) helped me analye the data's alignment to regression conditions.
