==> Final_Project/05_Analysis_and_Conclusions.Rmd <==
---
title: "Home Values Analysis and Conclusions"
author: "Ken Popkin"
---

```{r}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```

```{r echo=TRUE, warning=FALSE, message=FALSE}
library(dplyr)
library(ggplot2)
library(tidyverse)
library(reshape2)
library(ggpubr)
library(ggfortify)
library(psych)

require(scales)
```

### Introduction to Analysis
<I> Can education and wages be used as predictors for home values? </I>    
Now that Cleaning & Transformation and Visualization are complete, in this third notebook we'll complete a regression analysis to see how well home values can be predicted at the county level using education and wages. 

### About the Data
```{r echo=FALSE}
# load data
master <- read.csv('https://raw.githubusercontent.com/krpopkin/DATA607-Class-Repository/master/Final_Project/Data/master.csv')

cases = dim(master)[1]
features = dim(master)[2]
cat('Data for this project is', cases, 'cases and', features, 'features')
```

```{r data}
head(master,1)
```

```{r}
describe(master)
```

### Regression Analysis
The correlation heat maps and scatter plots in the previous Visualization notebook make it appear likely that all the features below will contribute to the regression model's accuracy.  Based on this we'll start with a full regression model for home values that uses all these features...        
   
region = census region (Northeast, South, Midwest, West)       
hs = high school diploma       
college = four year degree      
wages = average annual pay      
rc = resident count      
hv = home values      

<B> Create the Regression Model </B>  
```{r}
rmdata <- master %>%
  select(region_num, hs = hs_diploma_percent, college = four_year_degree_percent, wages = avg_annual_pay, rc = resident_count, hv = avg_home_value)

rmall <- lm(hv ~ region_num + hs + college + wages + rc, data = rmdata)
summary(rmall)
```

<B> Chart the regression model </B> <br>
We used five features in the regression model, but graphing all five would make for poor readability and isn't very practical. The chart below uses the features with the highest coefficients (wages and college) to visualize the regression line.
```{r}
ggplot(rmdata,aes(y=hv,x=wages,color=college)) + 
  geom_point() + 
  scale_x_continuous(labels = comma) +
  stat_smooth(method="lm",se=FALSE)
```
<br> One concern this chart raises is the nonlinear nature of the data as average annual wages start to exceed $60,000.  The data starts takes on a non-linear trend after this point and the overall distribution of the data is heteroschedastic, rather than the preferred homoschedastic. We'll look into this further when we evaluate the regression for reasonableness below.  

<B> Check for Reasonable Conditions </B>  
This model looks pretty good as all the p-values are a lot less than 0.05 and the adjusted R-squared is at 55.4%.  Prior to concluding that this is the final model, we'll check if the conditions to use Regression are reasonable.  To do this we will use the ggfortify library's autoplot function to create diagnostic plots.
```{r}
autoplot(rmall)
```

<B> Analyzing the regression relationships... </B>  
<I> Residuals vs Fitted </I> checks the linear relationship assumptions. Our results show that up to values of $500K the relationship is a horizontal line; above $500K the outliers take the data in an upward curve.

<I>Normal Q-Q </I> is used to examine whether the residuals are normally distributed. Similar to the "Residuals vs Fitted" the data skews right among outliers of high home values.

<I>Scale-Location (or Spread-Location) </I> is used to check the homogeneity of variance of the residuals (homoscedasticity). Our "$500K" pattern continues as there is good homoscedasticity up to this amount.

<I>Residuals vs Leverage </I> is used to identify influential cases, that is extreme values that might influence the regression results when included or excluded from the analysis. The dashed line in the plot is called "Cook's line" and values above this line should be considered for removal from the model.  The above result aligns with the prior notes that values above $500K appear to be impacting the model's ability to forecast with regression.

### Analysis Conclusions
Conclusions from this Analysis are...      
1. at a county level across the United States the explanatory variables of high school diploma, four year degree, wages, resident count, and region produce a useable regression model to forecast home values.<br> <br>
2. the plot of the regression model showed a heteroschedastic trend to the data for wages above $75,000.  In a similar manner the residuals plot for home values is homoschedastic up to around $500,000, then tends to be non-linear.<br><br>
3. to address the non-linear and heteroschedastic nature of this data for counties with higher average annual wages (above $75K) and/or home values above $500K a next step for this project could be to run this notebook excluding counties that meet these conditions.  It's reasonable to assume that the model's forecasting ability would improve.   

### References
[This article from Keong-Woong Moon](https://cran.r-project.org/web/packages/ggiraphExtra/vignettes/ggPredict.html) was helpful with graphing the linear regression line.
